+++ src/main/java/com/momega/spacesimulator/service/StaticTrajectoryManager.java
@@ -4,13 +4,15 @@ import com.momega.spacesimulator.model.MovingObject;
 import com.momega.spacesimulator.model.StaticTrajectory;
 import com.momega.spacesimulator.model.Trajectory;
 
+import java.math.BigDecimal;
+
 /**
  * Created by martin on 5/21/14.
  */
 public class StaticTrajectoryManager implements TrajectoryManager {
 
     @Override
-    public void computePosition(MovingObject movingObject, double newTimestamp) {
+    public void computePosition(MovingObject movingObject, BigDecimal newTimestamp) {
         StaticTrajectory trajectory = (StaticTrajectory) movingObject.getTrajectory();
         movingObject.setPosition(trajectory.getPosition());
     }
