+++ src/main/java/com/momega/spacesimulator/service/TrajectoryManager.java
@@ -4,6 +4,8 @@ import com.momega.spacesimulator.model.MovingObject;
 import com.momega.spacesimulator.model.Trajectory;
 import org.joda.time.DateTime;
 
+import java.math.BigDecimal;
+
 /**
  * Created by martin on 5/21/14.
  */
@@ -14,7 +16,7 @@ public interface TrajectoryManager {
      * @param movingObject
      * @param newTimestamp new timestamp
      */
-    void computePosition(MovingObject movingObject, double newTimestamp);
+    void computePosition(MovingObject movingObject, BigDecimal newTimestamp);
 
     /**
      * Indicates whether or not the service supports the trajectory
